bot/requirements.txt
python-telegram-bot==20.3
apscheduler==3.9.1.post1
python-dotenv==1.0.0
